refactor : 이메일 관련 dto Getter, 기본생성자 Lombok 추가
