--- src/main/java/com/prgrms/needit/common/domain/dto/IsUniqueResponse.java
@@ -1,10 +1,0 @@
-package com.prgrms.needit.common.domain.dto;
-
-public class IsUniqueResponse {
-
-	private final boolean isUnique;
-
-	public IsUniqueResponse(boolean isUnique) {
-		this.isUnique = isUnique;
-	}
-}
