+++ src/main/java/com/prgrms/needit/domain/user/login/dto/IsUniqueRequest.java
@@ -1,11 +1,13 @@
-package com.prgrms.needit.common.domain.dto;
+package com.prgrms.needit.domain.user.login.dto;
 
 import javax.validation.constraints.NotBlank;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 public class IsUniqueRequest {
 
 	@Getter
+	@NoArgsConstructor
 	public static class Email {
 
 		@NotBlank
@@ -17,6 +19,7 @@ public class IsUniqueRequest {
 	}
 
 	@Getter
+	@NoArgsConstructor
 	public static class Nickname {
 
 		@NotBlank
